refactor: componentize contest problem list label and button
